Pull out non chart specific functions into Helpers module
