+++ frontend/src/Monitor/Charts/Util.js
@@ -1,6 +1,7 @@
 import React from 'react';
 
 import {Line, Area} from 'recharts';
+import {shuffle, objectKeys} from 'Helpers'
 
 const NumberSuffixes = [
     {suffix: 'b', value: 1000000000},
@@ -124,19 +125,6 @@ function percentValueFormat(percent) {
     return `${percent.toFixed(2)}%`;
 }
 
-function objectKeys(object) {
-    return Object.entries(object).map(([key, _]) => key);
-}
-
-function shuffle(array) {
-    let shuffled = [...array];
-    for (let i = shuffled.length - 1; i > 0; i--) {
-        const j = Math.floor(Math.random() * (i + 1));
-        [shuffled[i], shuffled[j]] = [shuffled[j], shuffled[i]];
-    }
-    return shuffled;
-}
-
 function findKeysOverSeries(data, topKey) {
     let keys = {};
     data.forEach(snap => {
@@ -150,14 +138,6 @@ function findKeysOverSeries(data, topKey) {
 }
 
 const makeId = label => label.replace(/[\s()]/g, '');
-
-function capitalizeWord(word) {
-    return word.charAt(0).toUpperCase() + word.slice(1);
-}
-
-function capitalize(str, sep) {
-    return sep ? str.split(sep).map(capitalizeWord).join(' ') : capitalizeWord(str);
-}
 
 function renderLine(line, labelColor, onClick, strokeWidth=1) {
     return (
@@ -199,15 +179,6 @@ function renderArea(area, labelColor, stackId, onClick) {
     );
 }
 
-function randomString(length) {
-    let chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890';
-    let chosen = [];
-    for (let i = 0; i < length; i += 1) {
-        chosen.push(chars.charAt(Math.floor(Math.random()*chars.length)));
-    }
-    return chosen.join('');
-}
-
 export {
     dateTickFormat,
     valueTickFormat,
@@ -215,13 +186,8 @@ export {
     selectNested,
     findNested,
     getDataColors,
-    objectKeys,
     findKeysOverSeries,
-    shuffle,
     renderLine,
     renderArea,
-    makeId,
-    capitalizeWord,
-    capitalize,
-    randomString
+    makeId
 };
