+++ frontend/src/Monitor/Visualizations/Construction/ConstructionQueues.js
@@ -2,7 +2,8 @@ import React from 'react';
 
 import Chart from 'Monitor/Charts/Chart';
 import AreaChart from 'Monitor/Charts/AreaChart';
-import {selectNested, findKeysOverSeries, capitalize} from '../../Charts/Util';
+import {selectNested, findKeysOverSeries} from 'Monitor/Charts/Util';
+import {capitalize} from 'Helpers';
 import {registerChart} from 'Monitor/Charts/ChartRegistry';
 import {translate} from 'Translator';
