+++ frontend/src/Monitor/Visualizations/Economy/Gdp.js
@@ -8,7 +8,7 @@ import {makeStyles} from '@material-ui/core/styles';
 import Chart from 'Monitor/Charts/Chart';
 import LineChart from 'Monitor/Charts/LineChart';
 import {registerChart} from 'Monitor/Charts/ChartRegistry';
-import {selectNested} from '../../Charts/Util';
+import {selectNested} from 'Monitor/Charts/Util';
 import {translate} from 'Translator';
 
 const Name = 'Total Economy Value';
