+++ frontend/src/Monitor/Visualizations/Military/Armies.js
@@ -3,7 +3,7 @@ import React from 'react';
 import Chart from 'Monitor/Charts/Chart';
 import AreaChart from 'Monitor/Charts/AreaChart';
 import {registerChart} from 'Monitor/Charts/ChartRegistry';
-import {findKeysOverSeries, selectNested} from '../../Charts/Util';
+import {findKeysOverSeries, selectNested} from 'Monitor/Charts/Util';
 import {translate} from 'Translator';
 
 const Name = 'Army Breakdown';
