+++ frontend/src/Monitor/Visualizations/Science/Exploration.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import Chart from 'Monitor/Charts/Chart';
 import LineChart from 'Monitor/Charts/LineChart';
-import {selectNested} from '../../Charts/Util';
+import {selectNested} from 'Monitor/Charts/Util';
 import {registerChart} from 'Monitor/Charts/ChartRegistry';
 import {translate} from 'Translator';
