+++ frontend/src/Translator.js
@@ -5,7 +5,7 @@ import Select from '@material-ui/core/Select';
 import MenuItem from '@material-ui/core/MenuItem';
 import {makeStyles} from '@material-ui/core/styles';
 
-import {objectKeys} from './Monitor/Charts/Util';
+import {objectKeys} from 'Helpers';
 
 const useStyles = makeStyles((theme) => ({
     formControl: {
